Implement combineReducers logic, update imports
