+++ reducers/decks.js
@@ -1,0 +1,20 @@
+import { RECEIVE_DECKS, ADD_DECK } from "../actions/decks";
+
+function decks(state = {}, action) {
+  switch (action.type) {
+    case RECEIVE_DECKS:
+      return {
+        ...state,
+        ...action.decks
+      };
+    case ADD_DECK:
+      return {
+        ...state,
+        [action.deck.key]: action.deck
+      };
+    default:
+      return state;
+  }
+}
+
+export default decks;
