+++ reducers/index.js
@@ -1,20 +1,6 @@
-import { RECEIVE_DECKS, ADD_DECK } from "../actions/decks";
+import { combineReducers } from "redux";
+import decks from "./decks";
 
-function decks(state = {}, action) {
-  switch (action.type) {
-    case RECEIVE_DECKS:
-      return {
-        ...state,
-        ...action.decks
-      };
-    case ADD_DECK:
-      return {
-        ...state,
-        [action.deck.key]: action.deck
-      };
-    default:
-      return state;
-  }
-}
-
-export default decks;
+export default combineReducers({
+  decks
+});
